drop incorrect information about adding a tool
This repository contains only information about proprietary images.
None of the scripts to add a tool, build and test an image or even most
of the linked documents exist in this repository. Users can only request
a tool, but never add one themselves. There is no need for incomplete
documentation on an imposible process.